ytweb - comments load, collect refactor, remove chrome
