+++ App/SysExtensions/Net/HttpClientExtensions.cs
@@ -10,14 +10,13 @@ using System.Net.Http.Headers;
 using System.Text;
 using System.Threading.Tasks;
 using Humanizer;
+using Microsoft.Net.Http.Headers;
 using Newtonsoft.Json;
 using Polly;
 using Serilog;
 using SysExtensions.Security;
 using SysExtensions.Serialization;
 using SysExtensions.Text;
-using System.Net.Http.Headers;
-using Microsoft.Net.Http.Headers;
 
 namespace SysExtensions.Net {
   public static class HttpClientExtensions {
@@ -111,7 +110,7 @@ namespace SysExtensions.Net {
     public static Task<HttpResponseMessage> SendAsyncWithLog(this HttpClient client, HttpRequestMessage request, ILogger log = null,
       HttpCompletionOption completion = HttpCompletionOption.ResponseHeadersRead) {
       try {
-        return InnerSendAsyncWithLog(client, null, request, completion, log);
+        return InnerSendAsyncWithLog(client, getRequest: null, request, completion, log);
       }
       catch (TaskCanceledException e) {
         throw
@@ -123,7 +122,7 @@ namespace SysExtensions.Net {
     public static async Task<HttpResponseMessage> SendAsyncWithLog(this HttpClient client, Func<HttpRequestMessage> request, ILogger log,
       HttpCompletionOption completion, AsyncPolicy<HttpResponseMessage> policy) {
       try {
-        return await policy.ExecuteAsync(() => InnerSendAsyncWithLog(client, null, request(), completion, log));
+        return await policy.ExecuteAsync(() => InnerSendAsyncWithLog(client, getRequest: null, request(), completion, log));
       }
       catch (TaskCanceledException e) {
         throw
@@ -135,7 +134,7 @@ namespace SysExtensions.Net {
     public static async Task<HttpResponseMessage> SendAsyncWithLog(this HttpClient client, Func<Task<HttpRequestMessage>> getRequest, ILogger log,
       HttpCompletionOption completion, AsyncPolicy<HttpResponseMessage> policy) {
       try {
-        return await policy.ExecuteAsync(() => InnerSendAsyncWithLog(client, getRequest, null, completion, log));
+        return await policy.ExecuteAsync(() => InnerSendAsyncWithLog(client, getRequest, request: null, completion, log));
       }
       catch (TaskCanceledException e) {
         throw
@@ -220,10 +219,10 @@ namespace SysExtensions.Net {
     public static string UrlEncode(this string url) => WebUtility.UrlEncode(url);
 
     public static string UrlDecode(this string url) => WebUtility.UrlDecode(url);
-    
-    public static (string Name, string Value)[] Cookies(this HttpResponseHeaders headers) => 
-      headers.TryGetValues("Set-Cookie", out var values) 
-        ? SetCookieHeaderValue.ParseList(values.ToList()).Select(cookie => (cookie.Name.Value, cookie.Value.Value)).ToArray() 
-        : Array.Empty<(string,string)>();
+
+    public static (string Name, string Value)[] Cookies(this HttpResponseHeaders headers) =>
+      headers.TryGetValues("Set-Cookie", out var values)
+        ? SetCookieHeaderValue.ParseList(values.ToList()).Select(cookie => (cookie.Name.Value, cookie.Value.Value)).ToArray()
+        : Array.Empty<(string, string)>();
   }
 }
