+++ App/YtFunctions/ApiBackend.cs
@@ -14,6 +14,7 @@ using SysExtensions.Serialization;
 using SysExtensions.Text;
 using SysExtensions.Threading;
 using YtReader;
+using YtReader.Yt;
 using ExecutionContext = Microsoft.Azure.WebJobs.ExecutionContext;
 using IMSLogger = Microsoft.Extensions.Logging.ILogger;
