+++ App/YtReader/YtUpdater.cs
@@ -12,6 +12,7 @@ using YtReader.BitChute;
 using YtReader.Rumble;
 using YtReader.Search;
 using YtReader.Store;
+using YtReader.Yt;
 
 namespace YtReader {
   public class YtUpdaterCfg {
